finder/requirements.txt
asgiref==3.6.0
asyncio==3.4.3
beautifulsoup4==4.12.0
blinker==1.6.2
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
coverage==7.3.2
cryptography==40.0.2
distlib==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-embed-video==1.4.9
django-js-asset==2.0.0
django-restframework==0.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.0
Flask==3.0.0
fontawesomefree==6.2.0
googlemaps==4.10.0
gravityai==0.1.4.post2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
pathlib==1.0.1
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.5.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.2
six==1.16.0
sklearn==0.0.post11
slugify==0.0.1
soupsieve==2.3.1
sqlparse==0.4.2
tabulate==0.8.10
tzdata==2023.3
urllib3==1.26.8
virtualenv==20.23.0
virtualenv-clone==0.5.7
websockets==10.4
Werkzeug==3.0.0
whitenoise==6.0.0
